Moved /Order api calls to only be on the checkout
