+++ src/components/CheckoutOrderDetails/CheckoutOrderDetails.tsx
@@ -1,17 +1,18 @@
 import { ReactElement } from 'react';
 import { useLocalOrder } from 'context/OrderContext';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faPlusSquare, faMinusSquare, faTrashCan } from '@fortawesome/free-solid-svg-icons';
+import { faTrashCan } from '@fortawesome/free-solid-svg-icons';
+import cn from 'classnames';
 
 export const CheckoutOrderDetails = (): ReactElement => {
-  const { localOrder, decreaseProductQuantityOrRemove, removeProduct, increaseProductQuantity } = useLocalOrder();
+  const { localOrder, removeProduct } = useLocalOrder();
 
   const iconClass = `border rounded
   bg-gray-200
   h-[22px] w-[22px]
-  px-1 py-1 `;
+  px-1 py-1`;
 
-  const binClass = iconClass + 'hover:text-red-700 focus:outline-none';
+  const binClass = cn(iconClass, 'hover:text-red-700 focus:outline-none');
 
   return (
     <div className="font-sans text-[16px]">
@@ -19,35 +20,17 @@ export const CheckoutOrderDetails = (): ReactElement => {
         {localOrder.lineItems?.length > 0 ? (
           localOrder.lineItems.map((lineItem) => (
             <div key={lineItem.product.id}>
-              <div className="mb-3 mt-3 grid grid-cols-11" key={lineItem.product.id}>
-                <div className="col-span-5">
-                  <span className="">{lineItem.product.name}</span>
+              <div className="flex" key={lineItem.product.id}>
+                <div className="min-w-[10%] text-center">
+                  <span className={iconClass}>{lineItem.quantity}</span>
                 </div>
-                <div className="col-span-3 flex items-center justify-center">
-                  <button
-                    onClick={() => {
-                      increaseProductQuantity(lineItem.product);
-                    }}
-                  >
-                    <FontAwesomeIcon icon={faPlusSquare} className={iconClass} />
-                  </button>
-                  <span className="mx-3 mb-1">{lineItem.quantity}</span>
-                  <button
-                    onClick={() => {
-                      decreaseProductQuantityOrRemove(lineItem.product);
-                    }}
-                  >
-                    {lineItem.quantity === 1 ? (
-                      <FontAwesomeIcon icon={faTrashCan} className={binClass} />
-                    ) : (
-                      <FontAwesomeIcon icon={faMinusSquare} className={iconClass} />
-                    )}
-                  </button>
+                <div className="grow">
+                  <span>{lineItem.product.name}</span>
                 </div>
-                <div className="col-span-2 ml-auto">
+                <div className="min-w-[10%]">
                   <span>£{(lineItem.subTotal * 1e2) / 1e2}</span>
                 </div>
-                <div className="col-span-1 ml-auto">
+                <div className="min-w-[10%]">
                   <button
                     onClick={() => {
                       removeProduct(lineItem.product);
